to do ratio

+++ app/Services/ToDoService.js
@@ -1,9 +1,11 @@
 import {
   ProxyState
 } from "../AppState.js";
+
 import {
   ToDo
 } from "../Models/ToDo.js";
+
 import {
   sandboxApi
 } from "./AxiosService.js";
@@ -13,7 +15,6 @@ import {
 class ToDoService {
 
   constructor() {}
-
   async createToDo(formData) {
     console.log(formData);
     let toDos = ProxyState.toDos
@@ -25,15 +26,23 @@ class ToDoService {
   }
 
   async getToDo() {
-    ProxyState.toDos
     const res = await sandboxApi.get('Tristan/todos')
-    ProxyState.toDos = res.data
+    const toDos = res.data.map(t => new ToDo(t))
+    ProxyState.toDos = toDos
+    let totalToDos = toDos.length
+    let done = toDos.filter(done => done.completed == true)
+    let totalCompleted = done.length
+    console.log(totalToDos + " : " + totalCompleted);
+    document.getElementById('completed').innerHTML = (totalToDos + " : " + totalCompleted)
   }
 
   async deleteToDo(id) {
-    let toDos = ProxyState.toDos
-    const res = await sandboxApi.delete('Tristan/todos/' + id)
-    toDos = toDos.filter(t => t.id != id)
+
+    if (window.confirm("delete to-do?")) {
+      let toDos = ProxyState.toDos
+      const res = await sandboxApi.delete('Tristan/todos/' + id)
+      toDos = toDos.filter(t => t.id != id)
+    }
   }
 
 
@@ -44,6 +53,16 @@ class ToDoService {
     const res = await sandboxApi.put('Tristan/todos/' + found.id, found)
     toDos = toDos
   }
+
+  // countToDos() {
+  //   let toDos = ProxyState.toDos
+  //   totalToDos = toDos.length
+  //   let done = toDos.filter(done => done.completed == true)
+  //   let totalCompleted = done.length
+  //   console.log(totalToDos + " : " + totalCompleted);
+  //   document.getElementById('completed').innerHTML = (totalToDos + " : " + totalCompleted)
+  // }
+
 
 }
 export const toDoService = new ToDoService();
